Rename RelationBuilder file to Relation
